% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePolicyListing.R
\name{adjustGrowthRates}
\alias{adjustGrowthRates}
\title{Adjust Length of Growth Rates}
\usage{
adjustGrowthRates(grate, len, recycle)
}
\arguments{
\item{grate}{A vector of annual growth rates}

\item{len}{The output length of the growth rate vector}

\item{recycle}{A boolean indicator of whether the values of grate are recycled
i.e., reused in order multiple times, or if the last element of grate is copied
enough times to extend the length of grate to len}
}
\description{
Adjust Length of Growth Rates
}

