% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePolicyListing.R
\name{allocatePoliciesToMonth}
\alias{allocatePoliciesToMonth}
\title{Allocate annual number of policies to months of year}
\usage{
allocatePoliciesToMonth(num, weights)
}
\arguments{
\item{num}{The total number of policies to create for the year}

\item{weights}{A numeric vector of length 12 which specifies the distribution
of policies written over the calendar year. Defaults is uniformly
distributed}
}
\value{
A numeric vector of length 12 which indicates the number of policies
  to be created in each month
}
\description{
Allocate annual number of policies to months of year
}

