% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePolicyListing.R
\name{generatePolicyList}
\alias{generatePolicyList}
\title{Genereate Policy Listing}
\usage{
generatePolicyList(num, start, end, growthRate, monthlyExposureWeights,
  policyTerm, policyPrefix)
}
\arguments{
\item{num}{Either the total number of policies in the calendar year in which
start appears or a vector of calendar year policy counts}

\item{start}{A date object representing the beginning of the period for
which a policy listing will be created}

\item{end}{A date object representing the end of the period for which a
policy listing will be created}

\item{growthRate}{This is the percentage at which the number of policies is
expected to growth year over year (or a vector of percentages). Recycled
as necessary to fill out period between start and end. Default is 0.00
(i.e. no change)}

\item{monthlyExposureWeights}{A vector indicating how policies are allocated
by month across the calendar year. Defaults}

\item{policyTerm}{This is the length of the policy term in months. Default
is 12}

\item{policyPrefix}{This is used the generate the policy id for the policy
listing, default is "POL"}
}
\description{
Genereate Policy Listing
}

